try-catchに変更

+++ api/controllers/userController.js
@@ -13,26 +13,31 @@ module.exports = {
       console.log(error)
       return res.json({
         status: 400,
+        msg: 'failed',
         data: []
       })
     }
   },
-
-  // 新しいユーザーを作成する。
+  // 新しくユーザーを作成する。
   create_user : async (req, res) => {
-    const user = new User()
-    user.name = req.body.name
-    await user.save()
-      .then(result => {
-        res.json({ message: "User created"})
+    try {
+      const user = new User()
+      user.name = req.body.name
+      await user.save()
+      return res.json({
+        status: 200,
+        msg: 'success',
+        data: user
       })
-      .catch(err => {
-        res.send(err)
-        console.error(err)
+    } catch (error) {
+      console.log(error)
+      return res.json({
+        status: 400,
+        msg: 'failed',
+        data: []
       })
+    }
   },
-
-
   // 特定のユーザーを取得する。
   find_user : async (req, res) => {
     try {
@@ -42,11 +47,11 @@ module.exports = {
         msg: 'success',
         data: user
       })
-      // res.status(200).send("success")
     } catch (error) {
       console.log(error)
       return res.json({
         status: 400,
+        msg: 'failed',
         data: []
       })
     }
@@ -54,30 +59,37 @@ module.exports = {
 
   // 特定のユーザーを更新する。
   update_user : async (req, res) => {
-    const result =  await User.findById(req.params.user_id)
-      .then(user => {
-        user.name = req.body.name
-        return user
+    try {
+      const user = await User.findOneAndUpdate({_id: req.params.user_id}, { $set: { name: req.body.name } }, { new: true })
+      return res.json({
+        status: 200,
+        msg: 'success',
+        data: user
       })
-      .catch(err => {
-        res.send(err)
+    } catch (error) {
+      console.log(error)
+      return res.json({
+        status: 400,
+        msg: 'failed',
+        data: []
       })
-      await result.save()
-      .then(result => {
-        res.json({ message: "User created"})
-      })
-      .catch(err => {
-        res.send(err)
-      })
+    }
   },
 
   // 特定のユーザーを削除する。
   delete_user : async (req, res) => {
-    await User.deleteOne({
-      _id: req.params.user_id
-    })
-      .exec()
-      .then(() => res.json({ message: "Successfully deleted"}))
-      .catch(err => res.send(err))
+    try {
+      const user = await User.deleteOne({ _id: req.params.user_id }).exec()
+      return res.json({
+        status: 200,
+        msg: 'success',
+      })
+    } catch (error) {
+      return res.json({
+        status: 400,
+        msg: 'failed',
+        data: []
+      })
+    }
   }
 }
